Flytter feilobjekt ut, retter på modal-tekst og endrer Feil-fil klassen
