+++ src/frontend/komponenter/Fagsak/Personlinje/Behandlingsmeny/SettBehandlingTilbakeTilFakta/useSettBehandlingTilbakeTilFakta.test.ts
@@ -2,7 +2,7 @@ import { useQueryClient } from '@tanstack/react-query';
 import { renderHook } from '@testing-library/react';
 
 import { useSettBehandlingTilbakeTilFakta } from './useSettBehandlingTilbakeTilFakta';
-import { Feil } from '../../../../../api/Feil';
+import { Feil } from '../../../../../api/Ffeil';
 import * as TogglesContext from '../../../../../context/TogglesContext';
 import { RessursStatus } from '../../../../../typer/ressurs';
